Include main thread in the thread pool
By changing the threadpool to have t-1 threads, and making tasks that
spawn on other threads also have the spawing thread (executer) be a task
executer, parallel operations can more easily be nested.

Earlier if a task on the threadpool wanted to spawn tasks itself, it
would block one of threadpool tasks. This avoids that.